modularized simulated_annealing
